requirements.txt
# Product Sustainability & Impact Assessment API
# Python 3.12+

fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.32.0
gunicorn>=21.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.26.0
